pyproject.toml
[project]
name = "airflow-boilerpalte"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.11"
dependencies = [
    "ansible>=10.7.0",
    "apache-airflow==2.9.1",
    "pytest>=9.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
